Refactor mock leader in nonce cache tests
